WIP: Exercise 2
- message passes through to toast
- toast can open/close
- pass through icon using variant
