+++ src/components/Toast/Toast.js
@@ -1,33 +1,27 @@
 import React from 'react';
 import {
-  AlertOctagon,
-  AlertTriangle,
-  CheckCircle,
-  Info,
   X,
 } from 'react-feather';
 
 import VisuallyHidden from '../VisuallyHidden';
-
 import styles from './Toast.module.css';
 
-const ICONS_BY_VARIANT = {
-  notice: Info,
-  warning: AlertTriangle,
-  success: CheckCircle,
-  error: AlertOctagon,
-};
 
-function Toast() {
+function Toast({className, icon: Icon, handleClose, children}) {
+
+  // I think it would be better if we could use a Type in typescript to limit
+  // the possible values here.
+  const styleVariant = styles[className] || styles.notice;
+
   return (
-    <div className={`${styles.toast} ${styles.notice}`}>
+    <div className={`${styles.toast} ${styleVariant}`}>
       <div className={styles.iconContainer}>
-        <Info size={24} />
+        <Icon size={24}/>
       </div>
       <p className={styles.content}>
-        16 photos have been uploaded
+        {children}
       </p>
-      <button className={styles.closeButton}>
+      <button className={styles.closeButton} onClick={handleClose}>
         <X size={24} />
         <VisuallyHidden>Dismiss message</VisuallyHidden>
       </button>
